requirements.txt
fastapi
pyjwt
passlib
bcrypt
python-multipart
uvicorn